Review feedback for https://github.com/openxla/stablehlo/pull/1518
See details in comments.
