+++ BUILD.bazel
@@ -349,6 +349,7 @@ cc_library(
     deps = [
         ":reference_sizes",
         "@llvm-project//llvm:Support",
+        "@llvm-project//mlir:IR",
         "@llvm-project//mlir:Support",
     ],
 )
@@ -392,21 +393,6 @@ cc_library(
         ":reference_tensor",
         "@llvm-project//llvm:Support",
         "@llvm-project//mlir:Support",
-    ],
-)
-
-cc_library(
-    name = "reference_sizes",
-    srcs = [
-        "stablehlo/reference/Sizes.cpp",
-    ],
-    hdrs = [
-        "stablehlo/reference/Sizes.h",
-    ],
-    strip_include_prefix = ".",
-    deps = [
-        "@llvm-project//llvm:Support",
-        "@llvm-project//mlir:IR",
     ],
 )
