+++ stablehlo/reference/Ops.cpp
@@ -170,29 +170,6 @@ Tensor evalSliceOp(const Tensor &operand, const Index &index) {
   return evalSliceOp(operand, start, limit, strides);
 }
 
-void iterateThroughWindow(const Sizes &windowDimensions,
-                          const Sizes &windowStrides, const Sizes &operandShape,
-                          const Sizes &paddingLow, const Index &sourceIndex,
-                          std::function<void(const Index &)> body) {
-  auto rank = operandShape.size();
-  for (auto windowIndexIt =
-           IndexSpaceIterator(windowDimensions, Sizes(rank, 0));
-       windowIndexIt != IndexSpaceIterator(windowDimensions, std::nullopt);
-       ++windowIndexIt) {
-    Index operandIndex(rank);
-    bool outOfBound = false;
-    for (size_t i = 0; i < rank; ++i) {
-      operandIndex[i] = sourceIndex[i] * windowStrides[i] +
-                        (*windowIndexIt)[i] - paddingLow[i];
-      if (operandIndex[i] < 0 || operandIndex[i] >= operandShape[i]) {
-        outOfBound = true;
-        break;
-      }
-    }
-    if (!outOfBound) body(operandIndex);
-  }
-}
-
 }  // namespace
 
 SmallVector<Tensor> eval(
@@ -582,20 +559,19 @@ SmallVector<Tensor> eval(
         windowStrides.assign(windowStridesAttr.getValues<int64_t>().begin(),
                              windowStridesAttr.getValues<int64_t>().end());
 
-      Sizes paddingLow(rank, 0), paddingHigh(rank, 0);
+      Sizes paddingLow(rank, 0);
       if (auto padding = selectAndScatterOp.getPadding()) {
         auto paddingOrErr = hlo::convertPaddingAttribute(padding, {});
         if (failed(paddingOrErr))
           report_fatal_error(invalidArgument("Invalid padding format found."));
         for (auto i = 0; i < static_cast<int64_t>(paddingOrErr->size()); ++i) {
           paddingLow[i] = (*paddingOrErr)[i].first;
-          paddingHigh[i] = (*paddingOrErr)[i].second;
         }
       }
 
-      auto result = evalSelectAndScatter(
+      auto result = evalSelectAndScatterOp(
           operand, source, initValue, windowDimensions, windowStrides,
-          paddingLow, paddingHigh, selectAndScatterOp.getSelect(),
+          paddingLow, selectAndScatterOp.getSelect(),
           selectAndScatterOp.getScatter(), scope, selectAndScatterOp.getType());
       scope.add(op.getResults(), {result});
     } else if (auto shiftLeftOp = dyn_cast<ShiftLeftOp>(op)) {
@@ -1475,26 +1451,31 @@ Tensor evalSelectOp(const Tensor &pred, const Tensor &onTrue,
   return result;
 }
 
-Tensor evalSelectAndScatter(const Tensor &operand, const Tensor &source,
-                            const Tensor &initValue,
-                            const Sizes &windowDimensions,
-                            const Sizes &windowStrides, const Sizes &paddingLow,
-                            const Sizes &paddingHigh, Region &select,
-                            Region &scatter, Scope &scope,
-                            ShapedType resultType) {
+Tensor evalSelectAndScatterOp(const Tensor &operand, const Tensor &source,
+                              const Tensor &initValue,
+                              const Sizes &windowDimensions,
+                              const Sizes &windowStrides,
+                              const Sizes &paddingLow, Region &select,
+                              Region &scatter, Scope &scope,
+                              ShapedType resultType) {
   Tensor result(resultType, initValue.get({}));
-  auto rank = operand.getRank();
 
-  for (auto sourceIt = IndexSpaceIterator(source.getShape(), Sizes(rank, 0));
-       sourceIt != IndexSpaceIterator(source.getShape(), std::nullopt);
+  for (auto sourceIt = source.index_begin(); sourceIt != source.index_end();
        ++sourceIt) {
     std::optional<Element> selectedVal;
     std::optional<Index> selectedIndex;
+    auto iterateThroughWindow = [&](std::function<void(const Index &)> body) {
+      for (auto windowIt = windowDimensions.index_begin();
+           windowIt != windowDimensions.index_end(); ++windowIt) {
+        auto operandIndex = *sourceIt * windowStrides + *windowIt - paddingLow;
+        if (!operandIndex.inBounds(operand.getShape())) continue;
+        body(operandIndex);
+      }
+    };
     iterateThroughWindow(
-        windowDimensions, windowStrides, operand.getShape(), paddingLow,
-        *sourceIt, [&](const Index &operandIndex) {
+        [&](const Index &operandIndex) {
           auto currVal = operand.get(operandIndex);
-          if (!selectedVal.has_value()) {
+          if (!selectedVal) {
             selectedVal = currVal;
             selectedIndex = operandIndex;
           }
@@ -1514,8 +1495,7 @@ Tensor evalSelectAndScatter(const Tensor &operand, const Tensor &source,
           }
         });
     iterateThroughWindow(
-        windowDimensions, windowStrides, operand.getShape(), paddingLow,
-        *sourceIt, [&](const Index &operandIndex) {
+        [&](const Index &operandIndex) {
           if (operandIndex == selectedIndex) {
             Tensor sourceValues(
                 RankedTensorType::get({2}, initValue.getElementType()));
